requirements.txt
absl-py==0.15.0
adal==1.2.2
amqp==5.0.6
applicationinsights==0.11.9
argh==0.26.2
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.26
azure-core==1.9.0
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==10.3.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.6.0
azureml==0.2.7
azureml-core==1.18.0.post3
azureml-telemetry==1.18.0
azureml-train==1.18.0
azureml-train-core==1.18.0.post1
azureml-train-restclients-hyperdrive==1.18.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
base32hex==1.0.2
beautifulsoup4==4.10.0
billiard==3.6.4.0
bitarray==2.2.1
blis==0.4.1
boto==2.49.0
boto3==1.14.23
botocore==1.17.23
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.4
catalogue==0.0.8
catboost==0.26
celery==5.0.5
Cerberus==1.3.2
certifi==2020.6.20
cffi==1.13.2
characteristic==14.3.0
chardet==3.0.4
chronometry==2020.11.12
clang==5.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colouration==2020.7.26
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
dateparser==0.7.6
decorator==4.4.2
deskew==0.10.3
dill==0.3.4
disk==2021.6.29
distro==1.6.0
dnspython==2.1.0
docker==4.4.0
docopt==0.6.2
docutils==0.15.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
eventlet==0.32.0
flatbuffers==1.12
func-timeout==4.3.5
fuzzywuzzy==0.18.0
gast==0.4.0
gensim==3.8.3
geographiclib==1.52
geopy==2.2.0
gevent==21.8.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.16
greenlet==1.1.2
grpcio==1.41.1
h5py==3.1.0
hyperlink==19.0.0
idna==2.8
imageio==2.9.0
importlib-metadata==0.23
imutils==0.5.3
incremental==17.5.0
ipython==7.29.0
isodate==0.6.0
jedi==0.18.0
jeepney==0.6.0
jmespath==0.10.0
joblib==0.14.0
jsonpickle==1.4.1
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
klein==19.6.0
kombu==5.0.2
kq==2.2.0
lightgbm==2.3.0
lxml==4.6.4
Markdown==3.3.4
matplotlib==3.3.2
matplotlib-inline==0.1.3
more-itertools==7.2.0
msrest==0.6.10
msrestazure==0.6.2
murmurhash==1.0.2
ndg-httpsclient==0.5.1
networkx==2.5
nltk==3.6.5
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.3.0
packaging==21.0
pandas==1.2.4
parso==0.8.2
pathspec==0.8.1
pathtools==0.1.2
pdf==2020.11.12
pdf2image==1.14.0
pdfCropMargins==1.0.5
pdfminer==20191125
pdfminer.six==20211012
pdfreader==0.1.10
pdfrw==0.4
pdftotree==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pipreqs==0.4.10
pkg-resources==0.0.0
plac==1.1.3
plotly==4.14.3
preshed==3.0.2
price-parser==0.3.1
prompt-toolkit==3.0.18
protobuf==3.19.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyenchant==3.2.2
pycryptodome==3.10.1
Pygments==2.10.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==3.11.3
PyMuPDF==1.19.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyPdf==1.13
PyPDF2==1.26.0
pytesseract==0.3.0
python-dateutil==2.8.1
python-Levenshtein==0.12.1
python3-ghostscript==0.5.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
ravenclaw==2020.11.12
rdflib==6.0.2
regex==2021.11.2
reportlab==3.5.67
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.21.3
scipy==1.3.2
SecretStorage==3.3.0
selectivesearch==0.4
Send2Trash==1.8.0
setuptools-scm==6.3.2
silverware==2020.11.12
six==1.13.0
sklearn==0.0
slytherin==2020.11.12
smart-open==2.1.0
soupsieve==2.3
spacy==2.2.3
SPARQLWrapper==1.8.5
srsly==0.2.0
tabula-py==2.3.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.1
tensorflow-estimator==2.6.0
termcolor==1.1.0
tesserocr==2.5.1
thinc==7.3.1
threadpoolctl==2.1.0
tifffile==2020.10.1
tomli==1.2.1
tqdm==4.38.0
traitlets==5.1.1
Tubes==0.2.0
Twisted==19.10.0
txt==2020.11.12
typing==3.7.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.7
vine==5.0.0
Wand==0.6.7
wasabi==0.4.0
watchdog==0.9.0
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==0.16.0
wrapt==1.12.1
XlsxWriter==1.3.7
xmltodict==0.12.0
yarg==0.1.9
zipp==0.6.0
zope.event==4.5.0
zope.interface==4.7.1
